requirements.txt
asgiref==3.4.1
aspose-words==22.7.0
async-generator==1.10
attrs==21.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
colorama==0.4.4
cryptography==35.0.0
decorator==4.4.2
deprecation==2.1.0
Django==3.2.8
django-crispy-forms==1.13.0
django-filter==21.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
docx2pdf==0.1.8
ffmpeg-python==0.2.0
fire==0.4.0
fonttools==4.34.4
free-proxy==1.0.3
future==0.18.2
h11==0.12.0
idna==3.3
imageio==2.16.0
imageio-ffmpeg==0.4.5
img2pdf==0.4.4
importlib-metadata==4.8.1
gunicorn==20.1.0
lxml==4.6.3
Markdown==3.3.4
moviepy==1.0.3
numpy==1.21.5
opencv-python==4.6.0.66
outcome==1.1.0
packaging==21.3
pdf2docx==0.5.5
pikepdf==5.4.0
Pillow==8.4.0
proglog==0.1.9
psycopg2-binary==2.9.3
pycparser==2.20
PyMuPDF==1.20.1
pyOpenSSL==21.0.0
pyparsing==3.0.9
python-docx==0.8.11
pytz==2021.3
pywin32==304
requests==2.26.0
selenium==4.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.2.1
sqlparse==0.4.2
termcolor==1.1.0
tqdm==4.62.3
trio==0.19.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.7
wsproto==1.0.0
zipp==3.6.0